Add more authentication in the servlets Added Course remind, Eval Remind, Eval Publish, and Eval Unpublish servlets Fixed the HTML for the change of header Clean up BrowserInstance Test: CoordCourseAddPageUiTest finished
--HG--
rename : src/test/resources/pages/coordListCourseByIdNew.html => src/test/resources/pages/coordCourseByIdNew.html
rename : src/test/resources/pages/coordListCourseByNameNew.html => src/test/resources/pages/coordCourseByNameNew.html
rename : src/test/resources/pages/coordCourseAddDeleteInit.html => src/test/resources/pages/coordCourseDeleteInit.html
rename : src/test/resources/pages/coordCourseAddDeleteSuccessful.html => src/test/resources/pages/coordCourseDeleteSuccessful.html
rename : src/test/resources/pages/coordListCourseEmpty.html => src/test/resources/pages/coordCourseEmpty.html

+++ src/main/java/teammates/servlet/CoordEvalDeleteServlet.java
@@ -6,6 +6,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import teammates.api.Common;
+import teammates.datatransfer.CourseData;
 import teammates.jsp.Helper;
 
 @SuppressWarnings("serial")
@@ -36,7 +37,13 @@ public class CoordEvalDeleteServlet extends ActionServlet<Helper> {
 		// Get parameters
 		String courseID = req.getParameter(Common.PARAM_COURSE_ID);
 		String evalName = req.getParameter(Common.PARAM_EVALUATION_NAME);
-		System.out.println(evalName);
+		CourseData course = helper.server.getCourse(courseID);
+		if(course!=null && !course.coord.equals(helper.userId)){
+			helper.statusMessage = "You are not authorized to delete the evaluation " +
+					Helper.escapeHTML(evalName)+" in course "+courseID;
+			helper.redirectUrl = Common.PAGE_COORD_EVAL;
+			return;
+		}
 		
 		// Process action
 		helper.server.deleteEvaluation(courseID,evalName);
